#300 - move commons.utils.enums package to commons.utils package
