+++ src/main/java/com/commercetools/sync/producttypes/utils/PlainEnumValueUpdateActionUtils.java
@@ -1,6 +1,6 @@
 package com.commercetools.sync.producttypes.utils;
 
-import com.commercetools.sync.commons.utils.enums.EnumValuesUpdateActionUtils;
+import com.commercetools.sync.commons.utils.EnumValuesUpdateActionUtils;
 import io.sphere.sdk.commands.UpdateAction;
 import io.sphere.sdk.models.EnumValue;
 import io.sphere.sdk.producttypes.ProductType;
@@ -12,9 +12,10 @@ import io.sphere.sdk.producttypes.commands.updateactions.RemoveEnumValues;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
+import java.util.Optional;
 
+import static com.commercetools.sync.commons.utils.CommonTypeUpdateActionUtils.buildUpdateAction;
 import static com.commercetools.sync.commons.utils.OptionalUtils.filterEmptyOptionals;
-import static com.commercetools.sync.commons.utils.enums.PlainEnumValueUpdateActionUtils.buildChangeLabelAction;
 
 public final class PlainEnumValueUpdateActionUtils {
 
@@ -68,12 +69,30 @@ public final class PlainEnumValueUpdateActionUtils {
             @Nonnull final EnumValue newEnumValue) {
 
         return filterEmptyOptionals(
-                buildChangeLabelAction(attributeDefinitionName,
-                        oldEnumValue,
-                        newEnumValue,
-                        ChangePlainEnumValueLabel::of
-                )
+                buildChangeLabelAction(attributeDefinitionName, oldEnumValue, newEnumValue)
         );
+    }
+
+    /**
+     * Compares the {@code label} values of an old {@link EnumValue} and a new {@link EnumValue}
+     * and returns an {@link Optional} of update action, which would contain the {@code "changeLabel"}
+     * {@link UpdateAction}. If both, old and new {@link EnumValue} have the same {@code label} values,
+     * then no update action is needed and empty optional will be returned.
+     *
+     * @param attributeDefinitionName the attribute definition name whose plain enum values belong to.
+     * @param oldEnumValue            the old plain enum value.
+     * @param newEnumValue            the new plain enum value which contains the new description.
+     * @return optional containing update action or empty optional if labels
+     *         are identical.
+     */
+    @Nonnull
+    public static Optional<UpdateAction<ProductType>> buildChangeLabelAction(
+        @Nonnull final String attributeDefinitionName,
+        @Nonnull final EnumValue oldEnumValue,
+        @Nonnull final EnumValue newEnumValue) {
+
+        return buildUpdateAction(oldEnumValue.getLabel(), newEnumValue.getLabel(),
+            () -> ChangePlainEnumValueLabel.of(attributeDefinitionName, newEnumValue));
     }
