+++ src/test/java/com/commercetools/sync/producttypes/utils/LocalizedEnumValueUpdateActionUtilsTest.java
@@ -1,4 +1,4 @@
-package com.commercetools.sync.commons.utils.enums;
+package com.commercetools.sync.producttypes.utils;
 
 import io.sphere.sdk.commands.UpdateAction;
 import io.sphere.sdk.models.LocalizedEnumValue;
@@ -10,7 +10,7 @@ import org.junit.Test;
 import java.util.Optional;
 
 
-import static com.commercetools.sync.commons.utils.enums.LocalizedEnumValueUpdateActionUtils.buildChangeLabelAction;
+import static com.commercetools.sync.producttypes.utils.LocalizedEnumValueUpdateActionUtils.buildChangeLabelAction;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class LocalizedEnumValueUpdateActionUtilsTest {
@@ -23,8 +23,7 @@ public class LocalizedEnumValueUpdateActionUtilsTest {
         final Optional<UpdateAction<ProductType>> result = buildChangeLabelAction(
             "attribute_definition_name_1",
             old,
-            newDifferent,
-            ChangeLocalizedEnumValueLabel::of
+            newDifferent
         );
 
         assertThat(result).contains(ChangeLocalizedEnumValueLabel.of("attribute_definition_name_1", newDifferent));
@@ -35,8 +34,7 @@ public class LocalizedEnumValueUpdateActionUtilsTest {
         final Optional<UpdateAction<ProductType>> result = buildChangeLabelAction(
             "attribute_definition_name_1",
             old,
-            newSame,
-            ChangeLocalizedEnumValueLabel::of
+            newSame
         );
 
         assertThat(result).isEmpty();
