+++ src/test/java/com/commercetools/sync/types/utils/typeactionutils/BuildLocalizedEnumUpdateActionsTest.java
@@ -10,10 +10,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import java.util.Collections;
 import java.util.List;
 
-import static com.commercetools.sync.commons.utils.enums.LocalizedEnumValueTestObjects.*;
+import static com.commercetools.sync.commons.utils.LocalizedEnumValueTestObjects.*;
 import static com.commercetools.sync.types.utils.LocalizedEnumValueUpdateActionUtils.buildLocalizedEnumValuesUpdateActions;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
