+++ src/test/java/com/commercetools/sync/types/utils/typeactionutils/BuildPlainEnumUpdateActionsTest.java
@@ -6,10 +6,9 @@ import io.sphere.sdk.types.commands.updateactions.AddEnumValue;
 import io.sphere.sdk.types.commands.updateactions.ChangeEnumValueOrder;
 import org.junit.Test;
 
-import java.util.Collections;
 import java.util.List;
 
-import static com.commercetools.sync.commons.utils.enums.PlainEnumValueTestObjects.*;
+import static com.commercetools.sync.commons.utils.PlainEnumValueTestObjects.*;
 import static com.commercetools.sync.types.utils.PlainEnumValueUpdateActionUtils.buildEnumValuesUpdateActions;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
